pyproject.toml
[project]
name = "sentinellayer"
version = "1.0.0"
description = "Network-native pre-authentication fraud intelligence for African fintech"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.18.4",
  "asyncpg>=0.31.0",
  "celery[redis]>=5.6.3",
  "confluent-kafka>=2.14.0",
  "fastapi>=0.136.1",
  "httpx>=0.28.1",
  "opentelemetry-exporter-otlp>=1.41.1",
  "opentelemetry-instrumentation-fastapi>=0.62b1",
  "opentelemetry-sdk>=1.41.1",
  "phonenumbers>=9.0.29",
  "prometheus-fastapi-instrumentator>=7.1.0",
  "pydantic-settings>=2.14.0",
  "redis>=6.4.0",
  "sentry-sdk[fastapi]>=2.58.0",
  "slowapi>=0.1.9",
  "sqlalchemy[asyncio]>=2.0.49",
  "structlog>=25.5.0",
  "tenacity>=9.1.4",
  "uvicorn[standard]>=0.46.0",
  "gunicorn>=22.0.0",
  "opentelemetry-distro>=0.47b0",
  "opentelemetry-instrumentation-sqlalchemy>=0.47b0",
  "opentelemetry-instrumentation-redis>=0.47b0",
  "firebase-admin>=7.4.0",
  "psycopg2-binary>=2.9.12",
  "google-genai>=1.74.0",
]

[dependency-groups]
dev = [
  "httpx>=0.28.1",
  "mypy>=1.20.2",
  "pre-commit>=4.6.0",
  "pytest>=9.0.3",
  "pytest-asyncio>=1.3.0",
  "pytest-cov>=7.1.0",
  "ruff>=0.15.12",
  "bandit>=1.7.9",
  "pytest-env>=1.6.0",
  "locust>=2.43.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "kafka"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "S", "B", "ANN"]
ignore = ["B008", "S101"]

[tool.mypy]
strict = true
python_version = "3.12"
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
